+++ XRContents8TeamProject/Assets/Scripts/PlayerScripts/PlayerManager.cs
@@ -18,6 +18,9 @@ public class PlayerManager : MonoBehaviour
     [Header("플레이어 스테이터스 관련")]
     [SerializeField] private float playerHp = 100.0f;
     [SerializeField] public float playerAtk = 10.0f;
+    [SerializeField] private float playerInvincibilityDuration = 1.0f;
+    private bool stateInvincibility = false;
+
 
     [Header("플레이어 사격 관련")]
     [SerializeField] public float shootSpeed = 1.0f;
